split out devicestate/handlers.go
